+++ ios/MullvadVPNUITests/SettingsMigrationTests.swift
@@ -138,7 +138,6 @@ class SettingsMigrationTests: BaseUITestCase {
             .tapUDPOverTCPPort80Cell()
             .tapQuantumResistantTunnelExpandButton()
             .tapQuantumResistantTunnelOnCell()
-            .tapMultihopSwitch()
     }
 
     func testVerifySettingsStillChanged() {
@@ -170,6 +169,5 @@ class SettingsMigrationTests: BaseUITestCase {
             .verifyUDPOverTCPPort80Selected()
             .tapQuantumResistantTunnelExpandButton()
             .verifyQuantumResistantTunnelOnSelected()
-            .verifyMultihopSwitchOn()
     }
 }